@@ -1,5 +1,5 @@
 
-var openid = require('openid');
+import openid from 'openid'
 
 import config from 'config/config'; // eslint-disable-line no-unused-vars
 import logger from 'app/logging'; // eslint-disable-line no-unused-vars
