@@ -1,42 +1,41 @@
-
-var request = require('request');
-var React = require('react');
-var ReactDOMServer = require('react-dom/server');
-var RecordPage = React.createFactory(require('public/client/js/react/build/record'));
-var MediaPage = React.createFactory(require('public/client/js/react/build/media'));
-var _ = require('lodash');
-
+import {createFactory} from '../createFactory.js'
+import request from 'request'
+import React from 'react'
+import ReactDOMServer from 'react-dom/server'
+import createRecordPage from 'public/client/js/react/build/record'
+import createMediaPage from 'public/client/js/react/build/media'
+import _ from 'lodash'
+// var RecordPage = createFactory(require('public/client/js/react/build/record'));
+// var MediaPage = createFactory(require('public/client/js/react/build/media'));
 import config from 'config/config'; // eslint-disable-line no-unused-vars
 import logger from 'app/logging'; // eslint-disable-line no-unused-vars
+const RecordPage = createFactory(createRecordPage);
+const MediaPage = createFactory(createMediaPage);
 
 export default {
   person: function(req, res) {
     res.render('person', {
       activemenu: 'people',
-      'uuid': req.params.id
+      uuid: req.params.id
     });
   },
-  // keep this for redirecting deprecated paths
   type: function(req, res) {
-    var t = req.params.type;
-    if(t === 'mediarecords' || t === 'records') {
-      res.redirect('/portal/' + t + '/' + req.params.id);
+    const t = req.params.type;
+    if (t === 'mediarecords' || t === 'records') {
+      res.redirect(`/portal/${t}/${req.params.id}`);
     } else {
-      res.send(404);
+      res.status(404).send('Not Found');
     }
   },
   record: function(req, res) {
-    var id = req.params.id;
-    // var Page = React.renderComponentToString(RecordPage({record: record, provider: recordset}));
-    request.get({"url": config.api + 'view/records/' + id, "json": true}, function(err, resp, body) {
-      if(err) {
+    const id = req.params.id;
+    request.get({ url: `${config.api}view/records/${id}`, json: true }, function(err, resp, body) {
+      if (err) {
         logger.error(err);
       }
-      if(body.uuid) {
-        var record = body;
-        var Page = ReactDOMServer.renderToString(RecordPage({record: record}));
-        //
-
+      if (body.uuid) {
+        const record = body;
+        let Page = ReactDOMServer.renderToString( <RecordPage record={record} /> );
         res.render('record', {
           activemenu: 'search',
           id: req.params.id,
@@ -47,8 +46,7 @@ export default {
           content: Page
         });
       } else {
-        res.status(404);
-        res.render('404', {
+        res.status(404).render('404', {
           activemenu: 'search',
           id: req.params.id,
           user: req.user,
@@ -57,35 +55,31 @@ export default {
       }
     });
   },
-
   media: function(req, res) {
-    // var qu = {size:1,from:0,query:{term:{uuid: id}}};
-    var id = req.params.id;
-
-    request.get({"url": config.api + 'view/mediarecords/' + id, "json": true}, function(err, resp, body) {
-      if(err) {
+    const id = req.params.id;
+    request.get({ url: `${config.api}view/mediarecords/${id}`, json: true }, function(err, resp, body) {
+      if (err) {
         logger.error(err);
       }
-
-      if(body.uuid) {
-        var mediarecord = body;
-        var record = {};
-        var render = function() {
-          var Page = ReactDOMServer.renderToString(MediaPage({mediarecord: mediarecord, record: record}));
+      if (body.uuid) {
+        const mediarecord = body;
+        let record = {};
+        const render = function() {
+          const Page = ReactDOMServer.renderToString(MediaPage({ mediarecord: mediarecord, record: record }));
           res.render('media', {
-              activemenu: 'search',
-              id: req.params.id,
-              user: req.user,
-              token: req.session._csrf,
-              content: Page,
-              record: record,
-              data: JSON.stringify({record: record, mediarecord: mediarecord}),
-              mediarecord: mediarecord
+            activemenu: 'search',
+            id: req.params.id,
+            user: req.user,
+            token: req.session._csrf,
+            content: Page,
+            record: record,
+            data: JSON.stringify({ record: record, mediarecord: mediarecord }),
+            mediarecord: mediarecord
           });
         };
-        if(_.has(mediarecord.indexTerms, 'records')) {
-          request.get({url: config.api + 'view/records/' + mediarecord.indexTerms.records[0], "json": true}, function(r_err, r_resp, r_body) {
-            if(r_err) {
+        if (_.has(mediarecord.indexTerms, 'records')) {
+          request.get({ url: `${config.api}view/records/${mediarecord.indexTerms.records[0]}`, json: true }, function(r_err, r_resp, r_body) {
+            if (r_err) {
               logger.error(r_err);
             }
             record = r_body;
@@ -95,35 +89,32 @@ export default {
           render();
         }
       } else {
-        res.status(404);
-              res.render('404', {
-                  activemenu: 'search',
-                  id: req.params.id,
-                  user: req.user,
-                  token: req.session._csrf
-              });
+        res.status(404).render('404', {
+          activemenu: 'search',
+          id: req.params.id,
+          user: req.user,
+          token: req.session._csrf
+        });
       }
-
     });
   },
-  // web crawlers specimen list for search engine indexing
   list: function(req, res) {
-    var base = 'https://s.idigbio.org/idigbio-list-pages/';
-    var page = 'index.html';
-    if(!_.isEmpty(req.params.page)) {
+    const base = 'https://s.idigbio.org/idigbio-list-pages/';
+    let page = 'index.html';
+    if (!_.isEmpty(req.params.page)) {
       page = req.params.page;
     }
-    request.get(base + page, function(err, resp, body) {
-      if(err) {
+    request.get(`${base}${page}`, function(err, resp, body) {
+      if (err) {
         logger.error(err);
       }
       res.render('list', {
-                activemenu: 'search',
-                id: req.params.id,
-                user: req.user,
-                token: req.session._csrf,
-                content: body
-            });
+        activemenu: 'search',
+        id: req.params.id,
+        user: req.user,
+        token: req.session._csrf,
+        content: body
+      });
     });
   }
 };
