@@ -1,5 +1,5 @@
-var React = require('react');
-var ReactDOM = require('react-dom');
-var CollectionPage = require('./react/src/collection');
+import React from 'react'
+import ReactDOM from 'react-dom'
+import CollectionPage from './react/src/collection'
 //var collection rendered in head of page
 ReactDOM.render(<CollectionPage data={collection} />, document.getElementById('collection-wrapper'));
