@@ -1,51 +1,58 @@
 
-var React = require('react');
-var dwc = require('../../lib/dwc_fields');
-var _ = require('lodash');
-var idbapi = require('../../lib/idbapi');
+import React, {useEffect, useState} from "react";
+import dwc from '../../lib/dwc_fields';
+import _ from 'lodash';
+import idbapi from '../../lib/idbapi';
+import Provider from './shared/provider';
+import Raw from './shared/raw';
+import Title from './shared/title';
 
-var Media = React.createClass({
-    errorImage: function(e){
+
+
+
+const Media = ({keyid, data}) => {
+    const [link, setLink] = useState('')
+    function errorImage(e){
         e.target.attributes['src'].value = '/portal/img/missing.svg';
-    },
-    render: function(){
-        var link='';
-        if(_.has(this.props.data,'ac:accessURI')){
-            link = this.props.data['ac:accessURI'];
-        }else if(_.has(this.props.data,'ac:bestQualityAccessURI')){
-            link = this.props.data['ac:bestQualityAccessURI'];
-        }else if(_.has(this.props.data, 'dcterms:identifier')){
-            link = this.props.data['dcterms:identifier'];
-        }
-
-        return (
-            <div key={this.props.keyid} id="media-wrapper" className="scrollspy section clearfix" >
-                <a className="clearfix" target={'_'+this.props.keyid} href={link} title="click to open original media file">
-                    <img className="media" 
-                    src={idbapi.media_host + 'v2/media/'+this.props.keyid+'?size=webview'} 
-                    onError={this.errorImage}/>
-                </a>
-                 Media retrieved from:<br />{link}
-                <a className="media-link hidden-print" href={link} target={'_'+this.props.keyid} title="click to open original media file">
-                    Open in browser<br />
-                </a>
-                <a className="media-link hidden-print" href={link} download={this.props.keyid} target={'_'+this.props.keyid} title="click to open original media file">
-                    Download File<br />
-                </a>
-            </div>
-        );
-
-    },
-    componentDidMount: function(){
-        
     }
-});
 
-var Table = React.createClass({
-    getInitialState: function(){
-        return {active: 'record'};
-    },
-    formatJSON: function(json){
+    useEffect(() => {
+        var link='';
+        if(_.has(data,'ac:accessURI')){
+            // link = data['ac:accessURI'];
+            setLink(data['ac:accessURI'])
+        }else if(_.has(data,'ac:bestQualityAccessURI')){
+            // link = data['ac:bestQualityAccessURI'];
+            setLink(data['ac:bestQualityAccessURI'])
+        }else if(_.has(data, 'dcterms:identifier')){
+            // link = data['dcterms:identifier'];
+            setLink(data['dcterms:identifier'])
+        }
+    }, []);
+
+    return (
+        <div key={keyid} id="media-wrapper" className="scrollspy section clearfix" >
+            <a className="clearfix" target={'_'+keyid} href={link} title="click to open original media file">
+                <img className="media"
+                src={idbapi.media_host + 'v2/media/'+keyid+'?size=webview'}
+                onError={errorImage}/>
+            </a>
+             Media retrieved from:<br />{link}
+            <a className="media-link hidden-print" href={link} target={'_'+keyid} title="click to open original media file">
+                Open in browser<br />
+            </a>
+            <a className="media-link hidden-print" href={link} download={keyid} target={'_'+keyid} title="click to open original media file">
+                Download File<br />
+            </a>
+        </div>
+    );
+
+};
+
+const Table = ({record}) => {
+    const [active, setActive] = useState('record')
+    const [rows, setRows] = useState([])
+    function formatJSON(json){
         if (typeof json != 'string') {
              json = JSON.stringify(json, undefined, 2);
         }
@@ -66,28 +73,32 @@ var Table = React.createClass({
             }
             return '<span class="' + cls + '">' + match + '</span>';
         });
-    },
-    tabClick: function(e){
+    }
+    function tabClick(e){
         e.preventDefault();
-        this.setState({active: e.target.attributes['data-tab'].value});
-    },
-    render: function(){
-        var order=[],rows=[],self=this;
+        setActive(e.target.attributes['data-tab'].value)
+        // this.setState({active: e.target.attributes['data-tab'].value});
+    }
+
+    useEffect(() => {
+
+        var order=[]
+        var temp_rows = []
 
         //make ordered name keys
         _.each(dwc.order.media,function(val){
-            if(_.has(self.props.record.data, val)){
+            if(_.has(record.data, val)){
                 order.push(val);
             }
         });
         //add unknown keys to end of list
-        var dif = _.difference(Object.keys(this.props.record.data),order);
+        var dif = _.difference(Object.keys(record.data),order);
         var merged = order.concat(dif), count=0;
         var regex = /(\bhttps?:\/\/(\S|\w| )+)/;
 
         _.each(order,function(key){
             var name = _.isUndefined(dwc.names[key]) ? key: dwc.names[key];
-            var val = self.props.record.data[key];
+            var val = record.data[key];
             if(_.isString(val)){
                 var str;
                 if(val.indexOf('<a')===0){
@@ -95,93 +106,106 @@ var Table = React.createClass({
                 }else{
                     str = val.replace(regex, "<a href=\"$1\">$1</a>");
                 }
-                rows.push(
+                temp_rows.push(
                     <tr key={count}>
                         <td className="name">{name}</td>
                         <td className="value" dangerouslySetInnerHTML={{__html: str}}></td>
                     </tr>
-                );                
+                );
             }else if(_.isArray(val)){
-                rows.push(
+                temp_rows.push(
                     <tr key={count}>
                         <td className="name">{name}</td>
                         <td className="value">{val.join(', ')}</td>
                     </tr>
-                );                 
+                );
             }else{
-                rows.push(
+                temp_rows.push(
                     <tr key={count}>
                         <td className="name">{name}</td>
                         <td className="value">{val}</td>
                     </tr>
-                );                 
+                );
             }
             count++;
         });
+        setRows([...rows, temp_rows])
+    }, []);
 
-        return (
-            <div id="data-table" className="scrollspy">
-                <ul className="tabs" onClick={this.tabClick}>
-                    <li className={this.state.active == 'record' ? 'active' : ''} data-tab="record">Data</li>
-                    <li className={this.state.active == 'raw' ? 'active' : ''} data-tab="raw">Raw</li>
-                </ul>
-                <section id="record" className="clearfix" style={{display: (this.state.active == 'record' ? 'block' : 'none' )}}>
-                    <table className="table table-striped table-condensed table-bordered">
-                    <tbody>{rows}</tbody>
-                    </table>
-                </section>
-                <section id="raw" style={{display: (this.state.active == 'raw' ? 'block' : 'none' )}}>
-                    <p id="raw-body" dangerouslySetInnerHTML={{__html: this.formatJSON(this.props.record)}}>
-                    </p>  
-                </section>  
-            </div>
-        );
-    }
-});
 
-var Group = React.createClass({
-    errorImage: function(e){
+    return (
+        <div id="data-table" className="scrollspy">
+            <ul className="tabs" onClick={tabClick}>
+                <li className={active == 'record' ? 'active' : ''} data-tab="record">Data</li>
+                <li className={active == 'raw' ? 'active' : ''} data-tab="raw">Raw</li>
+            </ul>
+            <section id="record" className="clearfix" style={{display: (active == 'record' ? 'block' : 'none' )}}>
+                <table className="table table-striped table-condensed table-bordered">
+                <tbody>{rows}</tbody>
+                </table>
+            </section>
+            <section id="raw" style={{display: (active == 'raw' ? 'block' : 'none' )}}>
+                <p id="raw-body" dangerouslySetInnerHTML={{__html: formatJSON(record)}}>
+                </p>
+            </section>
+        </div>
+    );
+
+};
+
+const Group = ({record, keyid}) => {
+    // constructor(props) {
+    //     super(props);
+    //     this.errorImage = this.errorImage.bind(this)
+    // }
+    function errorImage(e){
         e.target.attributes['src'].value = '/portal/img/missing.svg';
-    },
-    render: function(){
-        if( _.has(this.props.record,'indexTerms') && _.has(this.props.record.indexTerms,'mediarecords') && this.props.record.indexTerms.mediarecords.length > 1){
-            var imgs = [];
-            var media = this.props.record.indexTerms.mediarecords;
-            for(var id in media){
-                if(media[id] != this.props.keyid){
-                    imgs.push(
-                        <a href={'/portal/mediarecords/'+media[id]} title="click to open media record" key={media[id]} >
-                            <img className="gallery-image" src={idbapi.media_host + 'v2/media/'+media[id]+'?size=webview'} onError={this.errorImage} />
-                        </a>
-                    )                    
-                }
-            }
-            return (
-                <div id="other-images" className="clearfix scrollspy section">
-                    <h4 className="title">Other Media</h4>
-                    <div id="images-wrapper">
-                        {imgs}
-                    </div>
-                </div>
-            )
-        }else{
-            return null;
-        }
     }
-});
 
-var Provider = require('./shared/provider');
-var Raw = require('./shared/raw');
-var Title = require('./shared/title');
 
-module.exports = React.createClass({
-    taxaBreadCrumbs: function(){
+    if( _.has(record,'indexTerms') && _.has(record.indexTerms,'mediarecords') && record.indexTerms.mediarecords.length > 1){
+        var imgs = [];
+        var media = record.indexTerms.mediarecords;
+        for(var id in media){
+            if(media[id] != keyid){
+                imgs.push(
+                    <a href={'/portal/mediarecords/'+media[id]} title="click to open media record" key={media[id]} >
+                        <img className="gallery-image" src={idbapi.media_host + 'v2/media/'+media[id]+'?size=webview'} onError={errorImage} />
+                    </a>
+                )
+            }
+        }
+        return (
+            <div id="other-images" className="clearfix scrollspy section">
+                <h4 className="title">Other Media</h4>
+                <div id="images-wrapper">
+                    {imgs}
+                </div>
+            </div>
+        )
+    }else{
+        console.log('still working')
+        return null;
+    }
+
+};
+
+
+
+const MediaModExports = ({record, mediarecord}) => {
+    // constructor(props) {
+    //     super(props);
+    //
+    //     this.taxaBreadCrumbs = this.taxaBreadCrumbs.bind(this)
+    //     this.navList = this.navList.bind(this)
+    // }
+    function taxaBreadCrumbs(){
         var order = [], values = [], self = this;
         
         ['kingdom','phylum','class','order','family'].forEach(function(item){
-            if(_.has(self.props.record.indexTerms,item)){
+            if(_.has(record.indexTerms,item)){
                 order.push(item);
-                values.push(self.props.record.indexTerms[item]);
+                values.push(record.indexTerms[item]);
             }
         });
 
@@ -206,13 +230,13 @@ module.exports = React.createClass({
         });
 
         return output;
-    },
-    navList: function(){
+    }
+    function navList(){
 
         var othermedia = null;
         //var med = this.props.indexTerms.mediarecords
 
-        if( _.has(this.props.record,'indexTerms') && _.has(this.props.record.indexTerms,'mediarecords') && this.props.record.indexTerms.mediarecords.length > 1){
+        if( _.has(record,'indexTerms') && _.has(record.indexTerms,'mediarecords') && record.indexTerms.mediarecords.length > 1){
             othermedia = <li><a href="#other-images">Other Media</a></li>
         }
 
@@ -225,30 +249,31 @@ module.exports = React.createClass({
                 <li><a href="#data-table">All Data</a></li>
             </ul>  
         )
-    },
-    render: function(){
-        var source = this.props.mediarecord;
-        var info=[];
-        
-        return (
-            <div className="container-fluid">
-                <div className="row">
-                    <div className="col-lg-7 col-lg-offset-2 col-md-9 col-md-offset-1 col-sm-10" id="container">
-                        <h1 id="banner">Media Record</h1> 
-                        <span id="summary">{this.taxaBreadCrumbs()}</span>  
-                        <Title data={this.props.record} attribution={this.props.mediarecord.attribution} includeLink={true} mediaSearch={true} />
-                        <Media key={source.uuid+'_media'} keyid={source.uuid} data={source.data} />
-                        <Group record={this.props.record} keyid={source.uuid} />
-                        <Provider data={this.props.mediarecord.attribution} />
-                        <Table record={source} />
+    }
+
+    // var source = this.props.mediarecord;
+    var info=[];
+
+    return (
+        <div className="container-fluid">
+            <div className="row">
+                <div className="col-lg-7 col-lg-offset-2 col-md-9 col-md-offset-1 col-sm-10" id="container">
+                    <h1 id="banner">Media Record</h1>
+                    <span id="summary">{taxaBreadCrumbs()}</span>
+                    <Title data={record} attribution={mediarecord.attribution} includeLink={true} mediaSearch={true} />
+                    <Media key={mediarecord.uuid+'_media'} keyid={mediarecord.uuid} data={mediarecord.data} />
+                    <Group record={record} keyid={mediarecord.uuid} />
+                    <Provider data={mediarecord.attribution} />
+                    <Table record={mediarecord} />
+                </div>
+                <div className="col-lg-2 col-md-2 col-sm-2">
+                    <div id="side-nav">
+                        {navList()}
                     </div>
-                    <div className="col-lg-2 col-md-2 col-sm-2">
-                        <div id="side-nav">
-                            {this.navList()}
-                        </div>
-                    </div> 
                 </div>
             </div>
-        )
-    }
-})
+        </div>
+    )
+
+}
+export default MediaModExports;
