@@ -1,33 +1,41 @@
-
-var ReactDOM = require('react-dom');
-var React = require('react');
-//var Map = require('./search/views/mapbox');
-//window.queryBuilder = require('./search/lib/querybuilder');
-var async = require('async');
-var idbapi = require('./lib/idbapi');
-var L = require('leaflet');
+import ReactDOM from 'react-dom';
+import React from 'react';
+import async from 'async';
+import idbapi from './lib/idbapi';
+import L from 'leaflet';
 require('leaflet-sleep');
-//provides order for sections
-var RecordPage = require('./react/src/record');
-
+import RecordPage from './react/src/record'
+import MediaPage from "./react/src/media";
 
 var pubname = '';
-
-async.parallel([
-    function(callback){
-        idbapi.publishers({"pq":{"uuid":record.attribution.publisher}},function(resp){
-            resp.items.forEach(function(item){
-            pubname = item.data.name;
-            })
-            callback();
-        });
-    }
-],function(error){
-    ReactDOM.render(
-        <RecordPage record={record} pubname={pubname}/>,
-        document.getElementById('react-wrapper')
-    );
-})
+//Converted to csr due to ssr mismatch + ssr is not really needed here anyway.
+ReactDOM.render(
+    <RecordPage record={record} pubname={pubname}/>,
+    document.getElementById('react-wrapper')
+);
+// async.parallel([
+//     function(callback){
+//         idbapi.publishers({"pq":{"uuid":record.attribution.publisher}},function(resp){
+//             resp.items.forEach(function(item){
+//             pubname = item.data.name;
+//             })
+//             callback();
+//         });
+//     }
+// ],function(error){
+//     setTimeout(() => {
+//         ReactDOM.hydrate(
+//             <RecordPage record={record} pubname={pubname}/>,
+//             document.getElementById('react-wrapper')
+//         );
+//     }, [2000])
+//     ReactDOM.render(
+//         <RecordPage record={record} pubname={pubname}/>,
+//         document.getElementById('react-wrapper')
+//     );
+//
+//
+// })
 
 //$('.tabs .tab:first-child').trigger('click'); 
 //make map if geopoint
@@ -46,10 +54,10 @@ $('#side-nav-list').affix({
 
 if(_.has(record.indexTerms,'geopoint')){
     $('#map').css('display','block');
-   
+
     var base = L.tileLayer('//{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',{
         attribution: 'Map data © OpenStreetMap contributors',
-        minZoom: 0, 
+        minZoom: 0,
         maxZoom: 18,
         reuseTiles: true
     });
