@@ -1,6 +1,6 @@
-var React = require('react');
-var ReactDOM = require('react-dom');
-var StatsPage = require('./react/src/stats');
+import React from 'react'
+import ReactDOM from 'react-dom'
+import StatsPage from './react/src/stats'
 
 ReactDOM.render(
     <StatsPage usage={data.usage} ingest={data.ingest} ingestCumulative={data.ingestCumulative} collected={data.collected} taxon={data.taxon} flags={data.flags} />,
