@@ -1,65 +1,68 @@
+import React, { useState, useEffect, useRef } from 'react';
+import IDBMap from '../../../lib/mapper';
+import * as helpers from '../../../lib/helpers';
 
-var React = require('react');
-var IDBMap = require('../../../lib/mapper');
-var helpers = require('../../../lib/helpers');
+let map; // Declare map variable
 
-var map; 
-module.exports = React.createClass({
-    currentQuery: '',
+const Map = (props) => {
+    const [currentQuery, setCurrentQuery] = useState('');
+    const mapRef = useRef(null); // To store the map instance without triggering re-renders
 
-    componentDidMount: function(){
-        var self=this;
-        map = new IDBMap('map',{
-            queryChange: function(query){
-                var mapping;
-                if(_.has(query,'geopoint')){
-                    switch(query.geopoint.type){
-                        case 'geo_bounding_box':
-                            mapping={
-                                type: 'box',
-                                bounds: {
-                                    top_left: query.geopoint.top_left,
-                                    bottom_right: query.geopoint.bottom_right
-                                }
-                            }
-                            break;
-                        case 'geo_distance':
-                            mapping={
-                                type: 'radius',
-                                bounds: {
-                                    distance: query.geopoint.distance,
-                                    lat: query.geopoint.lat,//e.layer._latlng.lat,
-                                    lon: query.geopoint.lon //e.layer._latlng.lng
-                                }
-                            }
-                            break;
-                    } 
-                    self.props.viewChange('optionsTab','mapping');
-                    self.props.searchChange('mapping',mapping);                     
+    // Equivalent to componentDidMount and componentDidUpdate
+    useEffect(() => {
+        if (!mapRef.current) {
+            mapRef.current = new IDBMap('map', {
+                queryChange: function(query) {
+                    var mapping;
+                    if (_.has(query, 'geopoint')) {
+                        switch (query.geopoint.type) {
+                            case 'geo_bounding_box':
+                                mapping = {
+                                    type: 'box',
+                                    bounds: {
+                                        top_left: query.geopoint.top_left,
+                                        bottom_right: query.geopoint.bottom_right
+                                    }
+                                };
+                                break;
+                            case 'geo_distance':
+                                mapping = {
+                                    type: 'radius',
+                                    bounds: {
+                                        distance: query.geopoint.distance,
+                                        lat: query.geopoint.lat,
+                                        lon: query.geopoint.lon
+                                    }
+                                };
+                                break;
+                        }
+                        props.viewChange('optionsTab', 'mapping');
+                        props.searchChange('mapping', mapping);
+                    }
                 }
-            }
-        });
-
-        var query = queryBuilder.buildQueryShim(this.props.search);
-        map.query(query)
-    },
-    shouldComponentUpdate: function(){
-        return false;
-    },
-    componentWillReceiveProps: function(nextProps){
-        var q = queryBuilder.buildQueryShim(nextProps.search);
-        var next=JSON.stringify(q);
-        //debugger
-        if(next!==this.currentQuery){
-            this.currentQuery=next;
-            map.query(q);
+            });
         }
-    },
-    render: function(){
-        return (
-            <div id="map-wrapper">
-                <div id="map"></div>
-            </div>
-        )
-    }
-})
+
+        const query = queryBuilder.buildQueryShim(props.search);
+        mapRef.current.query(query);
+    }, [props.search]); // Only re-run if props.search changes
+
+    // Equivalent to UNSAFE_componentWillReceiveProps and shouldComponentUpdate
+    useEffect(() => {
+        const q = queryBuilder.buildQueryShim(props.search);
+        const next = JSON.stringify(q);
+
+        if (next !== currentQuery) {
+            setCurrentQuery(next);
+            mapRef.current.query(q);
+        }
+    }, [props.search, currentQuery]);
+
+    return (
+        <div id="map-wrapper">
+            <div id="map"></div>
+        </div>
+    );
+}
+
+export default Map;
